openaisdk-basedagents/requirements.txt
openai-agents
openai
python-dotenv